requested changes

+++ services/gateway/internal/model/snap.go
@@ -93,7 +93,7 @@ type RefreshSnap struct {
 	Download      *Download  `json:"download,omitempty"`
 	Version       *string    `json:"version,omitempty"`
 	Confinement   *string    `json:"confinement,omitempty"`
-	Revision      *int       `json:"revision,omitempty"`
+	Revision      *uint64    `json:"revision,omitempty"`
 	Type          *string    `json:"type,omitempty"`
 	Base          *string    `json:"base,omitempty"`
 }
@@ -155,3 +155,4 @@ type SnapPushRequest struct {
 type UnscannedUploadRequest struct {
 	Data string `json:"data"`
 }
+
